refactor(overlay): split the header button out and resync the wiki
Two rule violations raised across the PR's automated reviews.

`OverlayBoxHeaderButton` was a second top-level type sharing a file with
`OverlayBoxHeaderView`, against "Keep one primary type per file". It owns its
own rendering, hover tracking and symbol state, and `OverlayBoxPanel` reaches
it through the header, so it cannot be private: it moves to its own file.
`MovableLabel` stays where it is, being a private helper.

The wiki had drifted from the code it is supposed to be the source of truth
for. `settings-window.md` still named `viewDidEndLiveResize` as the drag hook,
which has been `OverlayBoxResizeAffordanceView.onResizeFinished` since the
affordance took over the drag, and it now also records that collapsing reports
the dragged height rather than the header's. `status.md` still inventoried only
the three pre-change overlay types.

No behaviour change.
